Organize files to include new round.js file and Round-test.js file
