+++ src/card.js
@@ -7,14 +7,6 @@ function createCard(id, question, answers, correctAnswer) {
   }
 }
 
-function evaluateGuess(guess, correctAnswer) {
-  if (guess === correctAnswer) {
-    return "correct!";
-  }
-
-  return "incorrect!";
-}
-
 function createDeck() {
   const deck  = [];
   const args = Array.from(arguments);
@@ -30,37 +22,8 @@ function countCards(deck) {
   return deck.length;
 }
 
-function createRound(deck) {
-  return {
-    deck: deck,
-    currentCard: deck[0],
-    turns: 0,
-    incorrectGuesses: [],
-    takeTurn: function(guess) {
-      const result = evaluateGuess(guess, this.currentCard.correctAnswer);
-
-      if(result === 'incorrect!'){
-        this.incorrectGuesses.push(this.currentCard.id);
-      }
-      
-      this.turns++;
-      this.currentCard = deck[this.turns];
-
-      return result;
-    },
-    calculatePercentageCorrect: function() {
-      return Math.round(100 - ((this.incorrectGuesses.length / this.turns) * 100));
-    },
-    endRound: function(){
-      return `** Round over! ** You answered ${this.calculatePercentageCorrect()}% of the questions correctly!`
-    }
-  }
-}
-
 module.exports = {
   createCard,
-  evaluateGuess,
   createDeck,
   countCards, 
-  createRound
 }
