+++ test/Card-test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-const { createCard, evaluateGuess, createDeck, countCards, createRound } = require('../src/card');
+const { createCard, createDeck, countCards} = require('../src/card');
 
 describe('card', function() {
   it('should be a function', function() {
@@ -50,127 +50,5 @@ describe('deck', function(){
     const numberOfCards = countCards(deck);
 
     expect(numberOfCards).to.equal(2);
-  })
-})
-
-describe('round', function() {
-  it('should be an function', function() {
-    expect(createRound).to.be.an('function');
-  })
-
-  it('should return a round and its properties', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-
-    const round = createRound(deck);
-
-    expect(round).to.be.an('object');
-    expect(round.deck).to.deep.equal(deck);
-    expect(round.currentCard).to.deep.equal(card1);
-    expect(round.turns).to.equal(0);
-    expect(round.incorrectGuesses).to.deep.equal([]);
-  })
-
-  it('should return a round with all of its methods', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-
-    const round = createRound(deck);
-
-    expect(round.takeTurn).to.be.a('function');
-    expect(round.calculatePercentageCorrect).to.be.a('function');
-    expect(round.endRound).to.be.a('function');
-  })
-})
-
-describe('turn', function() {
-  it('should update the turn count regardless of the result of the guess', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-    const round = createRound(deck);
-    const correctGuess = 'object';
-    const incorrectGuess = 'array';
-
-    round.takeTurn(correctGuess);
-    round.takeTurn(incorrectGuess);
-
-    expect(round.turns).to.equal(2);
-  })
-  
-  it('should update the currentCard to the next card', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-    const round = createRound(deck);
-
-    round.takeTurn('object', round);
-
-    expect(round.currentCard).to.deep.equal(card2);
-  })
-
-  it('should should add the cards id to incorrectGuesses if the guess is incorrect', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-    const round = createRound(deck);
-
-    const incorrectGuess = 'function';
-    const correctGuess = 'array';
-
-    round.takeTurn(incorrectGuess, round);
-    round.takeTurn(correctGuess, round);
-
-    expect(round.incorrectGuesses).to.deep.equal([1]);
-  })
-
-  it('should determine whether the guess is correct or not', function() {
-    const correctResult = evaluateGuess('object', 'object');
-    const incorrectResult = evaluateGuess('object', 'array');
-
-    expect(correctResult).to.equal('correct!');
-    expect(incorrectResult).to.equal('incorrect!');
-  })
-
-  it('should return feedback regardless of whether the guess is incorrect or not', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-    const round = createRound(deck);
-
-    const incorrectGuess = 'function';
-    const correctGuess = 'array'
-
-    incorrectFeedback = round.takeTurn(incorrectGuess, round);
-    correctFeedback = round.takeTurn(correctGuess, round);
-
-    expect(incorrectFeedback).to.equal('incorrect!');
-    expect(correctFeedback).to.equal('correct!');
-  })
-
-  it('should be able to calculate the percentage of correct guesses', function() {
-    const card1 = createCard(1, 'What allows you to define a set of related information using key-value pairs?', ['object', 'array', 'function'], 'object');
-    const card2 = createCard(2, "What is a comma-separated list of related values?", ["array", "object", "function"],"array")
-
-    const deck = createDeck(card1, card2);
-    const round = createRound(deck);
-
-    const incorrectGuess = 'function';
-    const correctGuess = 'array'
-
-    incorrectFeedback = round.takeTurn(incorrectGuess, round);
-    correctFeedback = round.takeTurn(correctGuess, round);
-
-    const percentage = round.calculatePercentageCorrect();
-
-    expect(percentage).to.equal(50);
   })
 })
